Add ForecastModels for hourly and daily forecast data
