+++ WeatherApp/WeatherApp/Core/ForecastModels.swift
@@ -1,0 +1,92 @@
+//
+//  ForecastModels.swift
+//  WeatherApp
+//
+//  Created by Phan Quyen on 07/08/2025.
+//
+
+import Foundation
+
+struct ForecastResponse: Codable {
+    let list: [ForecastItem]
+    let city: CityInfo
+}
+
+struct ForecastItem: Codable {
+    let dt: Int
+    let main: MainWeatherInfo
+    let weather: [Weather]
+    let dtTxt: String
+    
+    enum CodingKeys: String, CodingKey {
+        case dt, main, weather
+        case dtTxt = "dt_txt"
+    }
+}
+
+struct CityInfo: Codable {
+    let name: String
+    let country: String
+}
+
+struct HourlyForecast {
+    let time: Date
+    let temperature: Double
+    let description: String
+    let icon: String
+    
+    init(from item: ForecastItem) {
+        self.time = Date(timeIntervalSince1970: TimeInterval(item.dt))
+        self.temperature = item.main.temp
+        self.description = item.weather.first?.description.capitalized ?? ""
+        self.icon = item.weather.first?.icon ?? ""
+    }
+    
+    var timeString: String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "HH:mm"
+        return formatter.string(from: time)
+    }
+    
+    var temperatureString: String {
+        return String(format: "%.0f°", temperature)
+    }
+}
+
+struct DailyForecast {
+    let date: Date
+    let minTemperature: Double
+    let maxTemperature: Double
+    let description: String
+    let icon: String
+    
+    init(from items: [ForecastItem]) {
+        guard let firstItem = items.first else {
+            self.date = Date()
+            self.minTemperature = 0
+            self.maxTemperature = 0
+            self.description = ""
+            self.icon = ""
+            return
+        }
+        
+        self.date = Date(timeIntervalSince1970: TimeInterval(firstItem.dt))
+        
+        let temperatures = items.map { $0.main.temp }
+        self.minTemperature = temperatures.min() ?? 0
+        self.maxTemperature = temperatures.max() ?? 0
+        
+        self.description = firstItem.weather.first?.description.capitalized ?? ""
+        self.icon = firstItem.weather.first?.icon ?? ""
+    }
+    
+    var dayString: String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "EEEE"
+        return formatter.string(from: date)
+    }
+    
+    var temperatureRangeString: String {
+        return String(format: "%.0f° / %.0f°", minTemperature, maxTemperature)
+    }
+}
