+++ WeatherApp/WeatherApp/Core/WeatherModels.swift
@@ -58,28 +58,6 @@ struct SystemInfo: Codable {
     let sunset: Int
 }
 
-struct ForecastResponse: Codable {
-    let list: [ForecastItem]
-    let city: CityInfo
-}
-
-struct ForecastItem: Codable {
-    let dt: Int
-    let main: MainWeatherInfo
-    let weather: [Weather]
-    let dtTxt: String
-    
-    enum CodingKeys: String, CodingKey {
-        case dt, main, weather
-        case dtTxt = "dt_txt"
-    }
-}
-
-struct CityInfo: Codable {
-    let name: String
-    let country: String
-}
-
 struct WeatherData {
     let cityName: String
     let country: String
@@ -91,14 +69,14 @@ struct WeatherData {
     let timestamp: Date
     
     init(from response: WeatherResponse) {
-        cityName = response.name
-        country = response.sys.country
-        temperature = response.main.temp
-        description = response.weather.first?.description.capitalized ?? ""
-        icon = response.weather.first?.icon ?? ""
-        humidity = response.main.humidity
-        windSpeed = response.wind.speed
-        timestamp = Date(timeIntervalSince1970: TimeInterval(response.dt))
+        self.cityName = response.name
+        self.country = response.sys.country
+        self.temperature = response.main.temp
+        self.description = response.weather.first?.description.capitalized ?? ""
+        self.icon = response.weather.first?.icon ?? ""
+        self.humidity = response.main.humidity
+        self.windSpeed = response.wind.speed
+        self.timestamp = Date(timeIntervalSince1970: TimeInterval(response.dt))
     }
     
     var temperatureString: String {
@@ -111,65 +89,5 @@ struct WeatherData {
     
     var windString: String {
         return String(format: "%.1f m/s", windSpeed)
-    }
-}
-
-struct HourlyForecast {
-    let time: Date
-    let temperature: Double
-    let description: String
-    let icon: String
-    
-    init(from item: ForecastItem) {
-        time = Date(timeIntervalSince1970: TimeInterval(item.dt))
-        temperature = item.main.temp
-        description = item.weather.first?.description.capitalized ?? ""
-        icon = item.weather.first?.icon ?? ""
-    }
-    
-    var timeString: String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "HH:mm"
-        return formatter.string(from: time)
-    }
-    
-    var temperatureString: String {
-        return String(format: "%.0f°", temperature)
-    }
-}
-
-struct DailyForecast {
-    let date: Date
-    let minTemperature: Double
-    let maxTemperature: Double
-    let description: String
-    let icon: String
-    
-    init(from items: [ForecastItem]) {
-        guard let firstItem = items.first else {
-            date = Date()
-            minTemperature = 0
-            maxTemperature = 0
-            description = ""
-            icon = ""
-            return
-        }
-        
-        date = Date(timeIntervalSince1970: TimeInterval(firstItem.dt))
-        let temperatures = items.map { $0.main.temp }
-        minTemperature = temperatures.min() ?? 0
-        maxTemperature = temperatures.max() ?? 0
-        description = firstItem.weather.first?.description.capitalized ?? ""
-        icon = firstItem.weather.first?.icon ?? ""
-    }
-    
-    var dayString: String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "EEEE"
-        return formatter.string(from: date)
-    }
-    
-    var temperatureRangeString: String {
-        return String(format: "%.0f° / %.0f°", minTemperature, maxTemperature)
     }
 }
